[refactor] Introduce new library/ dir
It will look like:

    library/
        builtin_pure.py
        builtin_misc.py
        ...
        func_hay.py
        func_misc.py
