+++ library/func_misc.py
